+++ etl/src/v2/mod.rs
@@ -1,3 +1,4 @@
+pub mod concurrency;
 pub mod destination;
 pub mod pipeline;
 pub mod replication;
